Gallery api and tests

+++ server/controllers/galleries.js
@@ -8,41 +8,69 @@ galleryRouter.get('/', async (request, response) => {
 });
 galleryRouter.get('/:groupBy', async (request, response) => {
   const galleryItems = await Gallery.find({ groupBy: request.params.groupBy });
-  response.json(galleryItems);
+
+  if (galleryItems.length === 0) {
+    return response.status(404).json({ error: 'No items found or invalid path ' });
+  }
+  return response.json(galleryItems);
 });
 
 galleryRouter.post('/', async (request, response) => {
-  const { title, img, groupBy } = request.body;
-  const gallery = new Gallery({
-    title,
-    img,
-    groupBy,
-  });
+  if (!request.employee) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  if (request.employee.accessLevel === 1) {
+    const { title, img, groupBy } = request.body;
+    const gallery = new Gallery({
+      title,
+      img,
+      groupBy,
+    });
 
-  const savedToGallery = await gallery.save();
-  response.json(savedToGallery);
+    const savedToGallery = await gallery.save();
+    return response.status(201).json(savedToGallery);
+  }
+  return response
+    .status(403)
+    .json({ error: `Don't have permission to add a new gallery item` })
+    .end();
 });
 
 galleryRouter.put('/:id', async (request, response) => {
-  const { body } = request;
+  if (!request.employee) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  if (request.employee.accessLevel === 1) {
+    const { title, groupBy } = request.body;
 
-  const gallery = {
-    title: body.title,
-    groupBy: body.groupBy,
-  };
+    const gallery = {
+      title,
+      groupBy,
+    };
 
-  const updatedGalleryItem = await Gallery.findByIdAndUpdate(request.params.id, gallery, {
-    new: true,
-    runValidators: true,
-    context: 'query',
-  });
+    const updatedGalleryItem = await Gallery.findByIdAndUpdate(request.params.id, gallery, {
+      new: true,
+      runValidators: true,
+      context: 'query',
+    });
 
-  response.json(updatedGalleryItem);
+    return response.json(updatedGalleryItem);
+  }
+  return response.status(403).json({ error: `Don't have permission to update gallery item` }).end();
 });
 
 galleryRouter.delete('/:id', async (request, response) => {
-  await Gallery.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  if (!request.employee) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  if (request.employee.accessLevel === 1) {
+    const deletedGalleryItem = await Gallery.findByIdAndRemove(request.params.id);
+    if (!deletedGalleryItem) {
+      return response.status(404).end();
+    }
+    return response.status(204).end();
+  }
+  return response.status(403).json({ error: `Don't have permission to delete gallery item` }).end();
 });
 
 module.exports = galleryRouter;
